Fixed bug in vme get_data() functions: used to be 16-bit return value, which is too small for v1190 TDCs which have 19 bits of data. Changed to int32_t return, so this should fix it.

+++ src/vme/V792.hxx
@@ -42,7 +42,7 @@ public:
 	void reset();
 
 	/// Get a data value, with bounds checking
-	int16_t get_data(int16_t ch) const
+	int32_t get_data(int16_t ch) const
 		{
 			/*!
 			 * \param ch Channel number to get data from
